[core] Mark ipc processes by their name
It is possible to debug the ipc processes started by the
`ipc-connection-provider`, but it might be hard to find the name.

This commit adds each server name to the command line used to start
them, allow us to use `--serverName-inpect` option more easily.

It also adds a bunch of debug configuration to be able to attach to
multiple processes at the same time. VS Code doesn't allow us to
relaunch the same configuration, hence the duplicates.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>